Section 2 (#2)
* Added login functionality.

* Added a signup endpoint.

* Added ability to upload picture and authentication check.

* Added Get user details API. Fixed password issue with signup API.

* Added an update user method.

* Added auth to todo APIs

+++ functions/api/todos.js
@@ -12,6 +12,7 @@ exports.postTodo = (request, response) => {
     const newTodoItem = {
         title: request.body.title,
         body: request.body.body,
+        username: request.user.username,
         createdAt: new Date().toISOString()
     }
 
@@ -37,6 +38,9 @@ exports.getTodo = (request, response) => {
             if (!doc.exists) {
                 return response.status(404).json({ error: 'Todo not found' });
             }
+            if (!doc.data().username !== request.user.username) {
+                return response.status(403).json({ error: "Unaruthorized" });
+            }
             let todoData = doc.data();
             todoData.todoID = doc.id;
             return response.json(todoData);
@@ -50,6 +54,7 @@ exports.getTodo = (request, response) => {
 exports.getAllTodos = (request, response) => {
     db
         .collection('todos')
+        .where('username', '==', request.user.username)
         .orderBy('createdAt', 'desc')
         .get()
         .then((data) => {
@@ -99,6 +104,9 @@ exports.deleteTodo = (request, response) => {
         .then((doc) => {
             if (!doc.exists) {
                 return response.status(404).json({ error: 'Todo not found' });
+            }
+            if (!doc.data().username !== request.user.username) {
+                return response.status(403).json({ error: "Unaruthorized" });
             }
             return document.delete();
         })
